fix(query-relationships): return canonical entity label for the queried side
Address CodeRabbit review on #1503.

- Actionable: the query handler used the raw `entity` input for the queried
  side of subject/object while the other side used the resolved node's canonical
  label. Querying "jane doe" for the node "Jane Doe" yielded a mismatched pair.
  Use entityNode.label consistently. Added a regression test (case-insensitive
  resolution makes the raw-vs-canonical gap observable). Patch-bumped tool.json.
- Nitpick: extracted the duplicated in-memory EntityMemory / ToolContext test
  scaffolding shared by the two relationship-tool tests into
  skills/memory/tools/_shared/test-helpers.ts (no tool.json → skill discovery
  skips it). NB: the same scaffolding recurs in ~14 memory tests; a repo-wide
  test factory is the fuller fix, left for a follow-up.
