+++ skills/memory/tools/query-relationships/tool.json
@@ -1,7 +1,7 @@
 {
   "name": "query-relationships",
   "description": "Query entity-to-entity relationships from the knowledge graph. Resolves the entity by name. Returns all stored relationships, optionally filtered by edge type. When multiple nodes share the same name, returns an ambiguous response with candidates so you can ask the user to clarify.",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "sensitivity": "normal",
   "action_risk": "none",
   "inputs": {
